refactor test db utilities to fixtures/db_utils
